Addressed first round of feedback: - Cleaned up VME device logging - Addressed feedback on code that I didn't write - Converted all cases of LOGMASKED(LOG_GENERAL to LOG - Gated spammy logging in NES memory-mapper handlers behind LOG_HIFREQ - Added LOG_UNHANDLED and enabled it by default for NES memory-mapper   handlers that need it - Addressed issues in xbox_usb and xbox_nv2a - Enabled LOG_NANS by default in cem3394.cpp - Gated certain potentially weighty code behind VERBOSE mask checks

+++ src/devices/cpu/m6809/m6809.cpp
@@ -456,7 +456,7 @@ uint32_t m6809_base_device::execute_input_lines() const noexcept
 
 void m6809_base_device::execute_set_input(int inputnum, int state)
 {
-	LOGMASKED(LOG_INTERRUPTS, "%s: inputnum=%s state=%d totalcycles=%d\n", machine().describe_context(), inputnum_string(inputnum), state, (int) attotime_to_clocks(machine().time()));
+	LOGMASKED(LOG_INTERRUPTS, "%s: inputnum=%s state=%d totalcycles=%d\n", machine().describe_context(), inputnum_string(inputnum), state, attotime_to_clocks(machine().time()));
 
 	switch(inputnum)
 	{
